more description in docstring, moved helper function to quantum_system
